+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented here.
 
+## [Unreleased]
+### Changed
+- Rebranded OpenSpec as MindScript; update docs and rename core types to MindScriptBase/MindScriptContext/MindScriptTurn with zMindScript* validators (update imports accordingly).
+
 ## [0.1.0] - 2025-08-28
 ### Added
 - Initial release of the OpenSpec white paper